Split large file into GUI and Preferences part.
